Relocate App.js to components folder
